remove TeamEventsDao test file and update format
