+++ backend/src/dao/TeamEventsDao.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
 import { TeamEvent, DBTeamEvent } from '../DataTypes';
-import { memberCollection, teamEventsCollection } from '../firebase';
+import { teamEventsCollection } from '../firebase';
 import { NotFoundError } from '../errors';
 
 export default class TeamEventsDao {
